requirements.txt
Crawl4AI
python-dotenv
pydantic
